requirements.txt
astroid==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.5
django-appconf==1.0.2
django-compressor==2.2
django-constance==2.0.0
django-haystack==2.6.1
django-haystack-elasticsearch==0.1.0
django-hitcount==1.2.3
django-jet==1.0.6
django-modeltranslation==0.12.1
django-picklefield==1.0.0
django-taggit==0.22.1
django-widget-tweaks==1.4.1
elasticsearch==5.4.0
google-api-python-client==1.4.1
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
isort==4.2.15
Jinja2==2.9.6
lazy-object-proxy==1.3.1
mailchimp3==2.0.15
MarkupSafe==1.0
mccabe==0.6.1
oauth2client==4.1.2
olefile==0.44
pep8==1.7.0
Pillow==4.2.1
psycopg2==2.7.3.1
pyasn1==0.3.6
pyasn1-modules==0.1.4
pylint==1.7.3
python-memcached==1.58
pytz==2017.2
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
